[SPOT-274][REFACTOR] Move Services to Application Package
