+++ src/main/java/com/example/spot/legacy/web/controller/SearchController.java
@@ -5,8 +5,8 @@ import com.example.spot.refactor.common.api.code.status.SuccessStatus;
 import com.example.spot.refactor.study.domain.enums.StudySortBy;
 import com.example.spot.refactor.study.domain.enums.ThemeType;
 import com.example.spot.refactor.common.security.utils.SecurityUtils;
-import com.example.spot.legacy.service.study.StudyCommandService;
-import com.example.spot.legacy.service.study.StudyQueryService;
+import com.example.spot.refactor.study.application.StudyCommandService;
+import com.example.spot.refactor.study.application.StudyQueryService;
 import com.example.spot.legacy.web.dto.search.SearchRequestStudyDTO;
 import com.example.spot.legacy.web.dto.search.SearchRequestStudyWithThemeDTO;
 import com.example.spot.legacy.web.dto.search.SearchResponseDTO.HotKeywordDTO;
