+++ src/main/java/com/example/spot/refactor/study/application/StudyPostCommandServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.spot.legacy.service.studypost;
+package com.example.spot.refactor.study.application;
 
 import com.example.spot.refactor.common.api.code.status.ErrorStatus;
 import com.example.spot.refactor.common.api.exception.handler.MemberHandler;
