+++ src/main/java/com/example/spot/refactor/study/presentation/controller/StudyController.java
@@ -3,8 +3,8 @@ package com.example.spot.refactor.study.presentation.controller;
 import com.example.spot.refactor.common.api.ApiResponse;
 import com.example.spot.refactor.common.api.code.status.SuccessStatus;
 import com.example.spot.refactor.common.security.utils.SecurityUtils;
-import com.example.spot.legacy.service.study.StudyCommandService;
-import com.example.spot.legacy.service.study.StudyQueryService;
+import com.example.spot.refactor.study.application.StudyCommandService;
+import com.example.spot.refactor.study.application.StudyQueryService;
 import com.example.spot.refactor.study.domain.validation.annotation.ExistStudy;
 import com.example.spot.refactor.study.presentation.dto.request.StudyJoinRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyRegisterRequestDTO;
