+++ src/main/java/com/example/spot/refactor/study/presentation/controller/StudyPostController.java
@@ -4,8 +4,8 @@ import com.example.spot.refactor.common.api.ApiResponse;
 import com.example.spot.refactor.common.api.code.status.SuccessStatus;
 import com.example.spot.refactor.study.domain.enums.StudyPostCategoryQuery;
 import com.example.spot.legacy.service.s3.S3ImageService;
-import com.example.spot.legacy.service.studypost.StudyPostCommandService;
-import com.example.spot.legacy.service.studypost.StudyPostQueryService;
+import com.example.spot.refactor.study.application.StudyPostCommandService;
+import com.example.spot.refactor.study.application.StudyPostQueryService;
 import com.example.spot.refactor.study.domain.validation.annotation.ExistStudy;
 import com.example.spot.refactor.study.domain.validation.annotation.ExistStudyPost;
 import com.example.spot.refactor.study.domain.validation.annotation.ExistStudyPostComment;
