+++ src/test/java/com/example/spot/service/studypost/StudyPostCommandServiceTest.java
@@ -14,7 +14,7 @@ import com.example.spot.refactor.member.domain.MemberRepository;
 import com.example.spot.refactor.study.domain.aggregate.studymember.StudyMemberRepository;
 import com.example.spot.legacy.repository.NotificationRepository;
 import com.example.spot.refactor.study.domain.repository.StudyRepository;
-import com.example.spot.legacy.service.studypost.StudyPostCommandServiceImpl;
+import com.example.spot.refactor.study.application.StudyPostCommandServiceImpl;
 import com.example.spot.legacy.service.s3.S3ImageService;
 import com.example.spot.refactor.study.presentation.dto.request.StudyPostCommentRequestDTO;
 import com.example.spot.refactor.study.presentation.dto.request.StudyPostRequestDTO;
